# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend for Stitchfolio.",
        "title": "Stitchfolio-backend API docs",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/sf/v1",
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/task": {
            "get": {
                "description": "Get all active tasks",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get all active tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responseModel.Task"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Saves an instance of Task",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Save Task",
                "parameters": [
                    {
                        "description": "task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.Task"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            }
        },
        "/task/{id}": {
            "get": {
                "description": "Get an instance of Task",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get a specific Task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an instance of Task",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Update Task",
                "parameters": [
                    {
                        "description": "task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.Task"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Task id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an instance of Task",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Delete Task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requestModel.Task": {
            "type": "object",
            "properties": {
                "assignedToId": {
                    "type": "integer"
                },
                "completedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isCompleted": {
                    "type": "boolean"
                },
                "priority": {
                    "type": "integer"
                },
                "reminderDate": {
                    "type": "string"
                },
                "status": {
                    "description": "PENDING, IN_PROGRESS, COMPLETED, CANCELLED",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.AuditFields": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.DataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "pageMetaData": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "pageMetaData": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.Task": {
            "type": "object",
            "properties": {
                "assignedToId": {
                    "type": "integer"
                },
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "completedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isCompleted": {
                    "type": "boolean"
                },
                "priority": {
                    "type": "integer"
                },
                "reminderDate": {
                    "type": "string"
                },
                "status": {
                    "description": "PENDING, IN_PROGRESS, COMPLETED, CANCELLED",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}